requirements.txt
aiofiles==25.1.0
azure-functions==1.24.0
azure-storage-blob[aio]==12.27.1
